CODE_pdfreader/requirements.txt
fastapi
uvicorn
pypdf
pillow
pytesseract
python-multipart
